Домашнее задание 15

+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class ItemDto {
-    private Long id;
+    private Integer id;
     private String name;
     private String description;
     private Boolean available;
-    private Long request;
+    private Integer request;
 }
